Types of same domain visible across resources without explicit import.
This is analoguous to Java classes within the same package.
Done in DmxImportedNamespaceAwareLocalScopeProvider and would have been
a one-liner, but scoping broke. Now working again with another way of
declaring default imported namespaces:
- Updated Xtext due to a suspected scoping bug and regenerated all
languages
- DmxScopeProvider is no longer a subclass of
DmxImportedNamespaceAwareLocalScopeProvider but of DelegateScopeProvider
(via AbstractDmxScopeProvider) which is the Xtext default.
- Improved PlantUML generator with regards to references to external
aggreages in same domain.
- Change a warning to an error in DimValidator
- Fixed MANIFEST.MF warnings due to missing package exports

DIM-to-SIM transformation still needs adaptation to multiple entities
per aggregate.