Refactor PUB diagram provides, rename table providers.
- Introduce new (generation) target enumeration for HTML and LaTex (and
more in the future)
- Instead of referring to the concrete diagram renderer (including the
file format it generates) for each figure, implement a preference-based
mechanism on the abstractions "vector" and "raster"
- Introduce IDiagramFileFormatPreference and provide a default
implementation for each generator target for "vector" and "raster"
representations. This enables generating SVG for HTML and EPS for LaTeX.
- Change reference to a renderer with a specific file format for each
figure to a reference to the diagramType only. The file format is then
derived via the preferences of each target. Rename in syntax to "type"
from "figure" and "table", respectively.
- Implement a validator (on save) ensuring that actual renderers exist
(i.e. have been registered) for the given targets.
- Rename FigureRenderer and TableRenderer to ProvidedDiagramType and to
ProvidedTableType, respectively, to reflect the changed semantics of
those classes.