Refactor DIM entity origin & SIM query handling
- Rename DIM entity 'origin' to 'nature'
- Replace origin concept values by 'autonomous' nature
- Fix SIM scope provider for deduction-source references
- Fix various issues in DIM and SIM validation in the context of queries
- Fix SIM deduction of query parameters to also handle explicitly added
parameters
- Make use of 'add' in SIM consistent:
a) if parent is created via a rule (grab, ditch, morph, fuse) then new
members not created via a rule must be added with 'add' keyword.
b) if parent is not created via a rule, then no rules can be used to
define its members (exception: types and aggregates of an information
model).