Misleading error message on PUB diagram type.
Fixes a NP in the PUB scope provider when configured diagram type was
not supported for the given diagram root. Also this lead to a misleading
error message on the diagram type ("reference not cannot be resolved").
The scope provider only returned the supported diagram types. Now
returns all the diagram types and a validation provides a meaningful
error message.